Update notifications to be easier to work with
[Finishes #138262095](https://www.pivotaltracker.com/story/show/138262095)
[Fixes #139491427](https://www.pivotaltracker.com/story/show/139491427)
[#120143319](https://www.pivotaltracker.com/story/show/120143319)
