+++ src/constants/action_types.js
@@ -123,6 +123,7 @@ export const GUI = {
   SET_LAST_DISCOVER_BEACON_VERSION: 'GUI.SET_LAST_DISCOVER_BEACON_VERSION',
   SET_LAST_FOLLOWING_BEACON_VERSION: 'GUI.SET_LAST_FOLLOWING_BEACON_VERSION',
   SET_LAST_STARRED_BEACON_VERSION: 'GUI.SET_LAST_STARRED_BEACON_VERSION',
+  SET_NOTIFICATION_SCROLL_Y: 'GUI.SET_NOTIFICATION_SCROLL_Y',
   SET_SCROLL: 'GUI.SET_SCROLL',
   SET_SIGNUP_MODAL_LAUNCHED: 'GUI.SET_SIGNUP_MODAL_LAUNCHED',
   SET_VIEWPORT_SIZE_ATTRIBUTES: 'GUI.SET_VIEWPORT_SIZE_ATTRIBUTES',
