+++ src/containers/UserContainer.js
@@ -91,7 +91,7 @@ class UserContainer extends Component {
     externalLinksList: PropTypes.object,
     followersCount: PropTypes.number.isRequired,
     followingCount: PropTypes.number.isRequired,
-    formattedShortBio: PropTypes.string.isRequired,
+    formattedShortBio: PropTypes.string,
     invitationAcceptedAt: PropTypes.string,
     invitationEmail: PropTypes.string,
     id: PropTypes.string,
@@ -127,6 +127,7 @@ class UserContainer extends Component {
     categories: null,
     className: null,
     externalLinksList: null,
+    formattedShortBio: null,
     id: null,
     invitationAcceptedAt: null,
     invitationEmail: null,
