+++ src/routes/notifications/index.js
@@ -1,4 +1,4 @@
-import Notifications from '../../containers/notifications/Notifications'
+import NotificationsContainer from '../../containers/NotificationsContainer'
 
 const TYPES = [
   'all',
@@ -12,13 +12,13 @@ const TYPES = [
 export default {
   path: 'notifications(/:type)',
   getComponents(location, cb) {
-    cb(null, Notifications)
+    cb(null, NotificationsContainer)
   },
-  onEnter(nextState, replaceState) {
+  onEnter(nextState, replace) {
     const type = nextState.params.type
     // redirect back to /notifications if type is unrecognized
-    if (type && TYPES.indexOf(type) === -1) {
-      replaceState(nextState, '/notifications')
+    if (type && !TYPES.includes(type)) {
+      replace({ state: nextState, pathname: '/notifications' })
     }
   },
 }
