+++ src/routes/wtf/index.js
@@ -1,6 +1,6 @@
 export default {
   path: 'wtf(/**/**)',
-  onEnter(nextState, replaceState, callback) {
+  onEnter(nextState, replace, callback) {
     if (callback) {
       document.location.href = ENV.AUTH_DOMAIN + nextState.location.pathname
     }
